moved search components
